requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.7
django-environ==0.4.5
django-extensions==2.2.5
entrypoints==0.3
et-xmlfile==1.0.1
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
openpyxl==3.0.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.1.0
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
Send2Trash==1.5.0
simple-salesforce==0.74.3
simplejson==3.17.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
